backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

python-dotenv>=1.0.0
pydantic>=2.7.0
typing-extensions>=4.8.0

websockets>=12.0
langsmith>=0.1.0
